Split board tests with marsBoard
